Move AddEditClusterModal into clusters page components directory
